Making primary code adjustments.
